Select surfaces and subsurfaces by reference group
At the moment these are SurfaceName, SubSurfName, AllShadingSurfNames.

This introduces a few new things that we should consider, such as
where the surface does not have coordinates. This is fine in a lot of
cases since the geometry is defined relative to another surface.

In fields like `Wall:Exterior` though, we'll need to think about how to
handle them.